# Removes `_fe_equal_var`, and unwanted `_fe_normalize_weak` calls (in tests)