Fixes #946 and #1061

Changes:
- removes unwanted `fe_normalize_weak` calls to the second argument of `fe_equal`
- removes `fe_equal_var`