@siv2r 2022-01-13
In `fe_sqrt`:
https://github.com/bitcoin-core/secp256k1/blob/a1102b12196ea27f44d6201de4d25926a2ae9640/src/field_impl.h#L132-L135
Here, the `secp256k1_fe_sqr` function ensures that `t1->magnitude = 1` but there is no guarantee for the magnitude of `a` to be one since it is passed into the function. Hence, some `fe_sqrt` tests are failing after adding magnitude checks inside `fe_equal`.

We could do `secp256k1_fe_normalize_weak(a)` but this changes the value of `a` and we don't want that right? Alternatively, we could call `fe_equal_var` instead of `fe_equal`. Is it necessary for `secp256k1_fe_sqrt` need to be constant time?