@peterdettman 2022-01-13
@siv2r Regarding `_fe_sqrt`, there is an implicit magnitude restriction on `a`, since it is used as the input to `_fe_mul` and `_fe_sqr` at the top of the method. That magnitude restriction is `<= 8`. 8 is also a kind of global implicit limit for inputs where not otherwise stated (but we should definitely be working towards fully documenting and validating all such assumptions).

I think it might be better to just say that both inputs to the equals methods can be magnitude <= 8 too (indeed tied to the mul/sqr limit whatever it is or changes to). There's no real obstacle to the implementation(s) (just change the third argument to _fe_negate to 8) and no performance hit AFAICT.