@peterdettman 2022-01-13
> Is it necessary for `secp256k1_fe_sqrt` to be of constant time?

It's only used in a single var-time caller, so technically it could be changed to sqrt_var, but preferably not.