@real-or-random 2022-01-13
Apparently `secp256k1_fe_equal` and `secp256k1_fe_equal_var` are exactly identical and so they both only require that `a` has magnitude 1. So we should just get rid of `secp256k1_fe_equal_var` and `secp256k1_fe_equal` should have the magnitue restriction on a documented.