@siv2r 2022-01-13
> It's only used in a single var-time caller

Oh, Okay. Then, creating a new function for `secp256k1_fe_sqrt_var` (while renaming the old one to `_fe_sqrt_const`) might be overkill.

> we should just document that `secp256k1_fe_equal` only requires the magnitude of `a` to be 1.

Yes, I think this might be a better approach since it keeps the functionality of both `fe_equal_var` and `fe_equal` similar. 

After this, I felt we needed to remove the `_fe_normalize_weak` calls before `_fe_equal`, but surprisingly `_fe_equal` is used rarely in the code. The only other place where it is used is:
https://github.com/bitcoin-core/secp256k1/blob/a1102b12196ea27f44d6201de4d25926a2ae9640/src/modules/extrakeys/tests_impl.h#L75-L77
Even here, the magnitude of `y` is two, so modifying the `_fe_sqrt` function (which I suggested) is a bad idea.