@real-or-random 2022-01-13
> https://github.com/bitcoin-core/secp256k1/blob/a1102b12196ea27f44d6201de4d25926a2ae9640/src/modules/extrakeys/tests_impl.h#L75-L77

Hehe ok, and here it's basically an oversight. This is just test code, so `fe_equal_var` could be used instead. I mean, not that it matters.