@siv2r 2022-01-14
addressed https://github.com/bitcoin-core/secp256k1/pull/1062#discussion_r778917127 and https://github.com/bitcoin-core/secp256k1/pull/1062#pullrequestreview-844742940 (detailed explanation in 08186e8).