@siv2r 2022-01-14
I also benchmarked the `tests.c` file (using `time ./tests`) since a lot of `secp256k1_fe_verify` calls were added in this PR. These were the results (on 64-bit, i7-8750H) for three iterations:
| branch | iter1 | iter2 | iter3 |
|-----------|------|-------|-------|
| master |2min 26.21sec|2min 26.44sec|2min 26.68sec|
| this PR|2min 26.99sec|2min 26.91sec|2min 26.83sec|