@real-or-random 2022-01-18
@peterdettman Good points. Do you think we should remove `fe_equal_var` entirely and maybe document in `fe_equal` that this is intentional?