@peterdettman 2022-01-19
> @peterdettman Good points. Do you think we should remove `fe_equal_var` entirely and maybe document in `fe_equal` that this is intentional?

Yeah, pretty much. Until there's a performance-sensitive caller for whom a != b is the expected result, I don't see any value in `_fe_equal_var`. If such a caller does turn up, then there should be a fast path for inequality that doesn't even need the subtraction, so we'd rewrite it anyway.