@real-or-random 2022-01-27
@siv2r Do you plan to also remove the `fe_equal_var` in a further commit? Or should we do this in a separate PR?

(If you want to add it here, I think it's really ok to add it on top of the existing commits, no need to rewrite the existing commits).