@siv2r 2022-01-27
Okay, I will remove `fe_equal_var` in a new commit (this PR).