@siv2r 2022-01-30
- Removed `fe_equal_var` and documented the reason in `fe_equal`
- Replaced the `fe_equal_var` calls with `fe_equal`