@siv2r 2023-08-15
Rebased this PR. #1066 took care of all the magnitude and `fe_verify` checks so, this PR currently does:
- removes unwanted `fe_normalize_weak` calls to the second argument of `fe_equal`
- removes `fe_equal_var`