@real-or-random 2023-08-15
Nice. :)

> #1066 took care of all the magnitude and `fe_verify` checks so, this PR currently does:

Can you update the PR title and the initial comment, so that they reflect the current status?