@real-or-random 2023-08-16
Needs rebase again, sorry, but should be trivial. :) 

Thanks for updating the PR title. Can you also update the [initial post/PR description](https://github.com/bitcoin-core/secp256k1/pull/1062#issue-1093823651) (e.g., replace the contents with https://github.com/bitcoin-core/secp256k1/pull/1062#issuecomment-1678833854 )? It's just cosmetic, but the text of will be added to the merge commit, so it makes sense to keep this updated.